backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
PyPDF2
python-dotenv